not found error

+++ app/controllers/users_controller.ts
@@ -6,8 +6,14 @@ export default class UsersController {
     return await User.all()
   }
 
-  async show({ params }: HttpContext) {
-    const user = await User.findOrFail(params.id)
+  async show({ params, response }: HttpContext) {
+    const user = await User.find(params.id)
+    if (!user) {
+      return response.safeStatus(404).send({
+        status: 404,
+        message: 'not found',
+      })
+    }
     return {
       id: user?.id,
       firstName: user?.firstName,
@@ -22,22 +28,38 @@ export default class UsersController {
     })
   }
 
-  async update({ params, request }: HttpContext) {
-    const user = await User.findOrFail(params.id)
+  async update({ params, request, response }: HttpContext) {
+    const user = await User.find(params.id)
+    if (!user) {
+      return response.safeStatus(404).send({
+        status: 404,
+        message: 'not found',
+      })
+    }
     user.firstName = request.body().firstName
     user.lastName = request.body().lastName
     return await user.save()
   }
 
-  async delete({ params }: HttpContext) {
-    const user = await User.findOrFail(params.id)
+  async delete({ params, response }: HttpContext) {
+    const user = await User.find(params.id)
+    if (!user) {
+      return response.safeStatus(404).send({
+        status: 404,
+        message: 'not found',
+      })
+    }
+
     try {
       await user.delete()
       return {
+        status: 200,
         message: 'User deleted successfully',
       }
     } catch (error) {
+      console.log(error)
       return {
+        status: 500,
         message: 'Failed to delete user',
       }
     }
